nba-predictions-scheduled-job/requirements.txt
nba_api>=1.4.1
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.1.0
python-dotenv>=1.0.0